exercises_service/pyproject.toml
[tool.poetry]
name = "exercises-service"
version = "0.1.0"
description = ""
authors = ["Kirill Frolov <ffrolov.kirill@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
loguru = "^0.7.3"
python-dotenv = "^1.1.1"
alembic = "^1.16.4"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
